requirements.txt
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
Flask==2.2.2
greenlet==2.0.2
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.1
redis==4.4.2
SQLAlchemy==2.0.1
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.0
